refactor: カメラ制御処理をeventHandlersから分離
